requirements.txt
django==5.2.1
djangorestframework==3.16.0
drf_spectacular==0.28.0